Re-architecture cycle processing
Cycles are now processed by an object injected into the cycle counter
(rf_state) object. The API has changed so as to support this new
paradigm (mainly, the library keeps state in a new type of object
and not in rf_state anymore).

This was done so that we can add new counting methods in the future
other than the matrix method. For example, we could simply pass
the cycles out through the library API.