Optimized searching for flow to merge with
Given some flow(R[n]).

If the unterminated flow (if any) which comes immediatley before it
(R[n-1]) would not merge with R[n] then no flow would merge with R[n].

The reason for this is simple, if R[n - 1] does not need to merge
with R[n] but is also unterminated, it implies that it extends overtop
of R[n] and will therefore catch any unterminated flows coming above
which might have otherwise merged with R[n].

With this commit, we exploit this fact so that we don't have to
check every unterminated flow to see if it would merge with R[n] but
instead we only check the flow coming immediatley before.