Change pop_cycles functions to fill cycle_list
Added the cycle_list structure type which is now used to pass around
lists of cycles.

The cycle_list struct has also been used to fix the behaviour of
lib_rflow_pop_cycles_replace_mem so that it is now actually possible
for the caller to supply their own memory for cycle lists.